glyphsページリファクタ
